updates from PR and moar tests

+++ src/realm/sync/changeset_parser.cpp
@@ -90,12 +90,14 @@ struct State {
 };
 
 struct UnreachableInstructionHandler : public InstructionHandler {
-    void set_intern_string(uint32_t, StringBufferRange) override
+    void set_intern_string(uint32_t, StringBufferRange)
+    override
     {
         REALM_UNREACHABLE();
     }
 
-    StringBufferRange add_string_range(StringData) override
+    StringBufferRange add_string_range(StringData)
+    override
     {
         REALM_UNREACHABLE();
     }
@@ -712,8 +714,7 @@ OwnedMixed parse_base64_encoded_primary_key(std::string_view str)
         case Type::UUID:
             return OwnedMixed{state.read_uuid()};
         default:
-            throw BadChangesetError(util::format("invalid primary key type %1", type));
-
+            throw BadChangesetError(util::format("invalid primary key type %1", static_cast<int>(type)));
     }
 }
